monitoring/requirements.txt
pika==1.3.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
